backend/requirements.txt
asgiref==3.2.10
django==2.2.16
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-extra-fields==3.4.0
drf-yasg==1.20.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
PyJWT==2.4.0
Pillow==9.2.0
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
requests==2.26.0
